feat(middleware): optimize session handling and improve pricing plans integration
- Enhanced the middleware to quickly redirect requests to the root path, bypassing unnecessary checks for improved performance.
- Removed the PricingPlans component from the landing directory and updated imports across the application for better modularity.
- Deleted the outdated pricing-plans component, streamlining the codebase and reducing redundancy.
- Updated the billing management component to reflect the new import structure, ensuring consistency in pricing plan usage.
