+++ app/[locale]/dashboard/components/calendar/desktop-calendar.tsx
@@ -15,7 +15,7 @@ import { useI18n, useCurrentLocale } from "@/locales/client"
 import { WeeklyModal } from "./weekly-modal"
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover"
 import { HourlyFinancialTimeline } from "../mindset/hourly-financial-timeline"
-import { ImportanceFilter } from "@/components/importance-filter"
+import { ImportanceFilter } from "@/app/[locale]/dashboard/components/importance-filter"
 import { CountryFilter } from "@/components/country-filter"
 import { useNewsFilterStore } from "@/store/filters/news-filter-store"
 import { useCalendarViewStore } from "@/store/calendar-view"
