+++ app/[locale]/dashboard/components/mindset/news-impact.tsx
@@ -9,7 +9,7 @@ import { X } from "lucide-react"
 import { Badge } from "@/components/ui/badge"
 import { HourlyFinancialTimeline } from "@/app/[locale]/dashboard/components/mindset/hourly-financial-timeline"
 import { Skeleton } from "@/components/ui/skeleton"
-import { ImportanceFilter } from "@/components/importance-filter"
+import { ImportanceFilter } from "@/app/[locale]/dashboard/components/importance-filter"
 import { useNewsFilterStore } from "@/store/filters/news-filter-store"
 import { CountryFilter } from "@/components/country-filter"
 import { useFinancialEventsStore } from "../../../../../store/financial-events-store"
