+++ app/api/ai/chat/tools/ask-for-location.ts
@@ -1,0 +1,8 @@
+import { tool } from "ai";
+import { z } from "zod";
+
+export const askForLocation = tool({
+    description:
+        'Get the user location. Always ask for confirmation before using this tool.',
+    parameters: z.object({}),
+})
